Add bst check for pokespots refactor BST lookup into helper class Also put other "pure" functions in class as well
